입문주차 완성

+++ routes/comments.js
@@ -1,25 +1,26 @@
 const express = require("express");
+const comment = require('../schemas/comment');
 const Comment = require("../schemas/comment");
 const router = express.Router();
 //댓글 조회
 router.get("/:postId", async (req, res) => {
 
-    try {
         const _id = req.params.postId;
         const comments = await Comment.find({postId: _id});
-        console.log(comments.length)
+
+        if (comments.length === 0) {
+            return res.json({message : "댓글이 존재하지 않습니다."})
+        }
+
         const resultList = [];
         for (const comment of comments) {
             resultList.push(
                 {commentId: comment._id, user: comment.user, content: comment.content, createdAt: comment.createdAt}
             );
         }res.json(resultList)
-    } catch (e) {
-      //  if (comments.length === 0) {
-            return message("댓글이 존재하지 않습니다.")
-       // }
-    }
-});
+
+    })
+
 
 /* 댓글 작성 */
 router.post("/:postId", async (req, res) => {
@@ -35,6 +36,13 @@ router.post("/:postId", async (req, res) => {
 router.put("/:commentId", async (req, res) => {
     const _id = req.params.commentId;
     const {user, password, content} = req.body;
+    const comment = await Comment.findOne({_id})
+    if(!password) {
+        return res.json({message:"비밀번호를 입력해주세요"})
+    }
+    if(password !== comment.password){
+        return res.json({message: "비밀번호가 틀립니다."})
+    }
 
     await Comment.updateOne({_id}, {$set: {user,password,content}})
 
@@ -45,7 +53,14 @@ router.put("/:commentId", async (req, res) => {
  //댓글 삭제
 router.delete("/:commentId", async (req, res) => {
     const _id = req.params.commentId;
-
+    const {password} = req.body;
+    const comment = await Comment.findOne({_id})
+    if(!password) {
+        return res.json({message:"비밀번호를 입력해주세요"})
+    }
+    if(password !== comment.password){
+        return res.json({message: "비밀번호가 틀립니다."})
+    }
     await Comment.deleteOne({_id});
 
     res.json({message: "댓글이 삭제되었습니다." })
